Define & use $_SERVER config variables for test context
We need to test our wrapper and module against a local test server,
and against a main test server, and against a staging server, etc.

The default config of the wrapper and modules is to call the production API.

But, in other contexts, we need to force the definition of the test API server.
This was working so-so until now:
 - $_SERVER config values is set in the test webserver configuration,
 - test scripts and module code check for those variables:
   - CASHWAY_TEST_ENVIRONMENT
   - TEST_SERVER_SCHEME
   - TEST_SERVER_HOST
   - TEST_SERVER_PORT
   - and some others (see phpunit.xml).