PS: drastically simplify payment option buttons
CW payment button do not follow exactly the same template as others
and break a few shops layout, especially in small screen layouts.

This simplifies code & layout a lot, maybe a bit too much
(CW option does not stand out as much as before, but...
should this page become a battlefield between options? don't think so).

We can improve on this now.