MA: reorganize Makefile for packaging/releasing an archive
Now you only use `make package` when developing.

`make release` ensures that:
 - environment is clean to make the package,
 - versions match between readme doc and source code,
 - sets package archive version,
 - package,
 - sign the archive.

Archive name as evolved and is built as:
 - name (cashway-magento)
 - version (1.0)
 - optional dev version (-dev.{branch}.{ref})
 - archival timestamp (+12345..)

It is a bit long but should help to spot archives.
The point is that a packaged archive goes through QA later,
which in turn blesses it if it passes all tests,
so that it can be publicly released.